// seatunnel-ui/src/views/task/components/node/types.ts
type RelationType = 'AND' | 'OR'
type ITaskType = TaskType
type ScanType = 1 | 2 | 3 | 4
type IDateType = 'hour' | 'day' | 'week' | 'month'

interface IOption {
  label: string
  value: string | number
}
// ...
interface IDependpendItem {
  depTaskCode?: number
  status?: 'SUCCESS' | 'FAILURE'
  definitionCodeOptions?: IOption[]
  depTaskCodeOptions?: IOption[]
  dateOptions?: IOption[]
  projectCode?: number
  definitionCode?: number
  cycle?: 'month' | 'week' | 'day' | 'hour'
  dateValue?: string
  cardValue?: string
}

interface IDependTask {
  condition?: string
  nextNode?: number
  relation?: RelationType
  dependItemList?: IDependpendItem[]
}